refactor: consolidate etrace modules into _etrace_algorithms package
Merge _etrace_vjp/, _etrace_algorithms.py, _etrace_graph_executor.py,
and _snn_algorithms/ into a single flat _etrace_algorithms/ subpackage.

Collision-renames to preserve module identity:
  _etrace_vjp/base.py           -> vjp_base.py
  _etrace_vjp/graph_executor.py -> vjp_graph_executor.py

Top-level public API unchanged; all imports under braintrace.* resolve
the same symbols. Full suite 1220 passed.
