+++ braintrace/__init__.py
@@ -20,6 +20,22 @@ from . import nn
 from ._etrace_algorithms import (
     ETraceAlgorithm,
     EligibilityTrace,
+    ETraceGraphExecutor,
+    ETraceVjpAlgorithm,
+    ETraceVjpGraphExecutor,
+    ParamDimVjpAlgorithm,
+    D_RTRL,
+    IODimVjpAlgorithm,
+    ES_D_RTRL,
+    pp_prop,
+    EProp,
+    OSTL,
+    OTPE,
+    OTTT,
+    OSTTP,
+    FixedRandomFeedback,
+    KappaFilter,
+    PresynapticTrace,
 )
 from ._etrace_compiler import (
     ETraceGraph,
@@ -39,7 +55,6 @@ from ._etrace_compiler import (
     DiagnosticKind,
     DiagnosticLevel,
 )
-from ._etrace_graph_executor import ETraceGraphExecutor
 from ._input_data import SingleStepData, MultiStepData
 from ._etrace_op import (
     ETPPrimitive,
@@ -53,26 +68,7 @@ from ._etrace_op import (
     register_primitive_spec,
     get_primitive_spec,
 )
-from ._etrace_vjp import (
-    ETraceVjpAlgorithm,
-    ETraceVjpGraphExecutor,
-    ParamDimVjpAlgorithm,
-    D_RTRL,
-    IODimVjpAlgorithm,
-    ES_D_RTRL,
-    pp_prop,
-)
 from ._grad_exponential import GradExpon
-from ._snn_algorithms import (
-    EProp,
-    OSTL,
-    OTPE,
-    OTTT,
-    OSTTP,
-    FixedRandomFeedback,
-    KappaFilter,
-    PresynapticTrace,
-)
 from ._legacy import (
     ConvOp,
     ElemWiseOp,
