+++ braintrace/_etrace_algorithms/_common.py
