+++ braintrace/_etrace_algorithms/d_rtrl.py
@@ -21,7 +21,7 @@ import jax
 import jax.numpy as jnp
 import saiunit as u
 
-from braintrace._etrace_algorithms import EligibilityTrace
+from .base import EligibilityTrace
 from braintrace._etrace_compiler import HiddenParamOpRelation, HiddenGroup
 from braintrace._etrace_op import (
     etp_elemwise_p,
@@ -44,7 +44,7 @@ from braintrace._typing import (
     HiddenGroupJacobian,
     dG_Weight,
 )
-from .base import ETraceVjpAlgorithm
+from .vjp_base import ETraceVjpAlgorithm
 from .misc import (
     _extract_leaf,
     _reset_state_in_a_dict,
