+++ braintrace/_etrace_algorithms/misc.py
