+++ braintrace/_etrace_algorithms/osttp.py
@@ -26,7 +26,7 @@ import brainstate
 import jax
 import jax.numpy as jnp
 
-from braintrace._etrace_vjp.d_rtrl import ParamDimVjpAlgorithm
+from .d_rtrl import ParamDimVjpAlgorithm
 
 
 __all__ = ['OSTTP']
