+++ braintrace/_etrace_algorithms/otpe.py
@@ -29,8 +29,8 @@ import jax
 import jax.numpy as jnp
 
 from braintrace._etrace_op import is_batched_primitive
-from braintrace._etrace_vjp.base import ETraceVjpAlgorithm
-from braintrace._etrace_vjp.misc import _route_grads_by_path, _update_dict
+from .vjp_base import ETraceVjpAlgorithm
+from .misc import _route_grads_by_path, _update_dict
 from braintrace._misc import etrace_df_key
 from ._common import _resolve_leak
