+++ braintrace/_etrace_algorithms/public_api_test.py
@@ -18,7 +18,7 @@ import unittest
 
 class TestPublicAPI(unittest.TestCase):
     def test_subpackage_exports(self):
-        import braintrace._snn_algorithms as pkg
+        import braintrace._etrace_algorithms as pkg
         for name in (
             'EProp', 'OSTL', 'OTPE', 'OTTT', 'OSTTP',
             'FixedRandomFeedback', 'KappaFilter', 'PresynapticTrace',
